skip unit test
Need to discover how to write unit tests with composition api.
The unit test for features should import a component such as Home that displays the state of the feature flag. But that component needs to have a context which has provided the router.
Deferring this research for a short time.
unit test