Explicit naming of axios override module
Reduces confusion when reading code.  Note that normally many code editors elide the imports so a reader only sees the name "axios" in the code and it is too easy to assume this is the vanilla axios. But, in fact, it is an override of axios.  This submission simply  uses a distinct naming convention when using this override.
This change also allows exposure of underlying axios module for easier jest mock ups.

A second change is to remove the ability to set the base url. The PPR system has no one base url. It has at least three (api, auth, pay).  To increase consistency require base urls in all calls and avoid system defaults.