Deploy PPR UI on bcregistry.ca
To get here we need to make several adjustments in many different locations. This commit strives to reduce the number of moving parts to one environment variable that is pushed into the build configurations and then into the running environment. This variable is used by vue.config.js which provides it to the application BASE_URL which shares it with the rest of the application.

As well this commit provides a standard way to process and apply the build and deploy configs.  This standardizes our default configuration (just run the process.sh script) and allows for experimentation (e.g. load from a private repo branch and create new openshift resources that don't touch our running app) and for a rapid change to a new domain or subpath (we expect to change from /cooperatives/ppr to /ppr sometime in the near term).