Auth does not work on local dev
As a developer it is essential to run the app locally. Yet auth is, by design, tied to the bcreg domain.  

Developers need to disable auth on their local machine.

This workaround introduces risk that someone may check in code with this flag set to true.

Reviewers .. please discuss options.