Fix the reactive loading problem
The feature flags are to be reactive yet they are constructed and dependent on the loading of the LD client that needs to happen before the flags are created.   I believe we're having these chicken and egg problems because the Vue Composition API is using Vue 2.0 methods and when Vue 3 is released these problems will go away. Hopefully.  Search for info on Vue 3 and proxying data.

Several lines include this:
// eslint-disable-next-line @typescript-eslint/explicit-function-return-type

We need this fix for demo tomorrow. It was easier to use these disables than to figure out what Typescript wanted.  Maybe a reviewer has an idea?