A preliminary PPR Prototype
For commit history see the prototype branch.
All content is isolated into the proto folder and has no dependencies on another outside of this folder.  This is a stand alone Vue app that mimics a PPR system entirely in memory in a client's browser.

See the root readme.
See the numerous TODO sections on the various web pages. For example /src/views/FinancingStatementViewTab.vue inside the proto-to-do component's content.